Extract viewState to recycler items mapping into a dedicated class
